Move all files into package: duke
