+++ src/main/java/duke/DeleteCommand.java
@@ -1,4 +1,5 @@
-import java.io.IOException;
+package duke;
+
 import java.util.ArrayList;
 
 /**
