+++ src/main/java/duke/MarkCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.util.ArrayList;
 
 /**
