+++ src/main/java/duke/TaskList.java
@@ -1,4 +1,5 @@
-import java.io.*;
+package duke;
+
 import java.util.ArrayList;
 
 public class TaskList {
