+++ src/main/java/duke/TodoTask.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class TodoTask extends Task {
     public TodoTask(String input) {
         super(input);
