+++ tasklist.txt
@@ -1,5 +1,2 @@
-[D][X]      (by: 1 Oct 2020)
-[D][X]      (by: 3 Dec 2020)
-[T][X]      3
-[D][X]      4 (by: 1 Jan 2021)
-[E][X]      5 (at: 5)
+[T][ ]  hi
+[T][ ]  hie
